fix(processing): run Whitebox WASM tools off the main thread
The WASI runner is a single synchronous `wasi.start()` with no yield
points, so running it on the main thread freezes the whole UI -- no
repaint, no input -- for as long as the tool takes. Dissolving the
290-polygon layer from #1977 takes ~60s, and a stall probe measured
58,573 ms without a single animation frame: the app looks hung.

wasm-convert.ts already routed its tiling calls to a one-shot Worker for
exactly this reason, and its worker script is generic. Lift that
machinery into `wasm-tool-runner.ts`, rename the script to
`wasm-tool.worker.ts` now that it serves both callers, and route
`runWhiteboxToolWasm` through it. One implementation, so the two paths
cannot drift.

This freeze predates the dissolve fix -- every WASM tool blocked the main
thread -- but a long-running Dissolve was unreachable until the grouping
field became typeable, so it surfaces there first.

Measured on the same run: max main-thread stall drops from 58,573 ms to
756 ms, and the output is unchanged at 12 features.
